requirements.txt
pandas
python-dotenv 